requirements.txt
requests~=2.28.2
beautifulsoup4~=4.11.2
python-dotenv~=0.21.1
aiogram~=2.25.1